企業削除APIを修正

+++ app/Services/CompanyService.php
@@ -170,28 +170,34 @@ class CompanyService extends Service
         }
     }
 
-    public function delete(int $userId, int $companyId): \Illuminate\Http\JsonResponse
+    public function delete(int $userId, int $companyId): \App\Models\Company|array
     {
         try {
-            $company = $this->companyRepository->findBy(['user_id' => $userId, 'company_id' => $companyId]);
+            $user = $this->userRepository->find($userId);
+            if ( $user === null ) {
+                Log::error( __METHOD__ . ": User not found. (user_id={$userId})" );
+                return $this->errorUserNotFound();
+            }
 
+            $company = $this->companyRepository->findBy(['user_id' => $userId, 'company_id' => $companyId]);
             if ( $company === null ) {
-                return response()->notFound();
+                Log::error( __METHOD__ . ": Company not found. (user_id={$userId}, company_id={$companyId})" );
+                return $this->errorCompanyNotFound();
             }
 
             $isSuccess = $this->companyRepository->delete($company);
 
             if ( ! $isSuccess ) {
-                throw new Exception( __METHOD__ . ": Failed delete company. (company_id={$companyId})");
+                throw new Exception( __METHOD__ . ": Failed delete company. (user_id={$userId}, company_id={$companyId})");
             }
 
-            return response()->ok($company);
+            return $company;
 
         } catch ( Exception $e ) {
             Log::error(__METHOD__);
             Log::error($e);
 
-            return response()->internalServerError();
+            return $this->errorInternalServerError();
         }
     }
 }
